+++ Assets/Mirror/Editor/Logging/LogLevelWindow.cs
@@ -7,7 +7,7 @@ namespace Mirror.EditorScripts.Logging
     public class LogLevelWindow : EditorWindow
     {
         [Header("Log Settings Asset")]
-        [SerializeField] LogSettings settings = null;
+        [SerializeField] public LogSettings settings;
 
         SerializedObject serializedObject;
         SerializedProperty settingsProp;
@@ -42,7 +42,7 @@ namespace Mirror.EditorScripts.Logging
 
                         if (settings == null)
                         {
-                            LogSettings newSettings = LogLevelsGUI.DrawCreateNewButton();
+                            LogSettings newSettings = LogLevelsGui.DrawCreateNewButton();
                             if (newSettings != null)
                             {
                                 settingsProp.objectReferenceValue = newSettings;
@@ -51,7 +51,7 @@ namespace Mirror.EditorScripts.Logging
                         }
                         else
                         {
-                            LogLevelsGUI.DrawLogFactoryDictionary(settings);
+                            LogLevelsGui.DrawLogFactoryDictionary(settings);
                         }
                     }
                 }