+++ Assets/Mirror/Runtime/NetworkServer.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Mirror.RemoteCalls;
+using Mirror.AsyncTcp;
 using UnityEngine;
 using UnityEngine.Events;
 
@@ -122,6 +123,28 @@ namespace Mirror
 
         // transport to use to accept connections
         public AsyncTransport transport;
+
+        /// <summary>
+        /// Called when the script gets added to an object. Useful for getting other needed scripts.
+        /// </summary>
+        private void OnValidate()
+        {
+            // add transport if there is none yet. makes upgrading easier.
+            if (transport == null)
+            {
+                // First try to get the transport.
+                transport = GetComponent<AsyncTransport>();
+                // was a transport added yet? if not, add one
+                if (transport == null)
+                {
+                    transport = gameObject.AddComponent<AsyncTcpTransport>();
+                    logger.Log("NetworkServer: added default Transport because there was none yet.");
+                }
+#if UNITY_EDITOR
+                UnityEditor.Undo.RecordObject(gameObject, "Added default Transport");
+#endif
+            }
+        }
 
         /// <summary>
         /// This shuts down the server and disconnects all clients.